requirements.txt
django>=1.10

# Django Packages
django-crispy-forms
django-auth-ldap>=1.2.6

# PostGres
psycopg2>=2.6.1

# for .env file
django-environ

# Chained Selects (e.g. Pick facility then the associated instrument)
django-smart-selects

#For deploy
Fabric3>=1.13.1
python-dotenv

# Only for DEV
# To have a nice ./manage.py shell
ipython

# Only for production
uwsgi